[ylj]:修改添加订单到数据库å·以及è

+++ back-end/src/main/java/com/twuc/shopping/entity/ProductEntity.java
@@ -1,11 +1,14 @@
 package com.twuc.shopping.entity;
 
+import com.fasterxml.jackson.annotation.JsonIgnore;
+import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
 import javax.persistence.*;
+import java.util.List;
 
 @Entity
 @Table(name = "product")
@@ -18,7 +21,8 @@ public class ProductEntity {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Integer id;
     private String name;
-    private Integer price;
+    private Double price;
     private String unit;
     private String img;
+
 }
